+++ AllaganLib.Interface/Grid/IRenderTable.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Numerics;
+using System.Threading.Tasks;
 using ImGuiNET;
 
 namespace AllaganLib.Interface.Grid;
@@ -21,18 +22,21 @@ public interface IRenderTable<TConfiguration, TData, TMessageBase> where TConfig
     public int? FreezeRows { get; set; }
 
     public int? SortColumn { get; set; }
-    
+
     public bool ShowFilterRow { get; set; }
-    
+    public bool HasFooter { get; set; }
+
     public Func<TData, List<TMessageBase>>? RightClickFunc { get; set; }
 
     public ImGuiSortDirection? SortDirection { get; set; }
 
     public List<TData> GetItems();
+    public Task<List<TData>> GetItemsAsync();
 
-    public List<TData> GetFilteredItems(TConfiguration configuration);
+    public List<TData> GetFilteredItems(TConfiguration configuration, bool isDirty = false);
 
     public void SaveToCsv(TConfiguration configuration, RenderTableCsvExportOptions exportOptions);
 
     public List<TMessageBase> Draw(TConfiguration configuration, Vector2 size, bool shouldDraw = true);
+    public List<TMessageBase> DrawFooter(TConfiguration configuration, Vector2 size, bool shouldDraw = true);
 }