+++ AllaganLib.Interface/Grid/IconColumn.cs
@@ -29,6 +29,8 @@ public abstract class IconColumn<TConfiguration, TData, TMessageBase> : IntegerF
 
     public abstract bool HideFilter { get; set; }
 
+    public virtual bool IsHidden { get; set; }
+
     public abstract ImGuiTableColumnFlags ColumnFlags { get; set; }
 
     public abstract string EmptyText { get; set; }
@@ -56,6 +58,12 @@ public abstract class IconColumn<TConfiguration, TData, TMessageBase> : IntegerF
         }
 
         return null;
+    }
+
+    public virtual List<TMessageBase> DrawFooter(TConfiguration config, List<TData> item, int columnIndex)
+    {
+        ImGui.TableNextColumn();
+        return new List<TMessageBase>();
     }
 
     public bool DrawFilter(TConfiguration configuration, IColumn<TConfiguration, TData, TMessageBase> column, int columnIndex)