+++ AllaganLib.Interface/Grid/StringColumn.cs
@@ -29,6 +29,8 @@ public abstract class StringColumn<TConfiguration, TData, TMessageBase> : String
 
     public abstract bool HideFilter { get; set; }
 
+    public virtual bool IsHidden { get; set; }
+
     public abstract ImGuiTableColumnFlags ColumnFlags { get; set; }
 
     public abstract string EmptyText { get; set; }
@@ -56,6 +58,12 @@ public abstract class StringColumn<TConfiguration, TData, TMessageBase> : String
     public bool DrawFilter(TConfiguration configuration, IColumn<TConfiguration, TData, TMessageBase> column, int columnIndex)
     {
         return this.stringColumnFilter.Draw(configuration, column, columnIndex);
+    }
+
+    public virtual List<TMessageBase> DrawFooter(TConfiguration config, List<TData> item, int columnIndex)
+    {
+        ImGui.TableNextColumn();
+        return new List<TMessageBase>();
     }
 
     public void SetupFilter(IColumn<TConfiguration, TData, TMessageBase> column, int columnIndex)